Introducing H2 for integration tests instead of PostgreSQL which needs a running server

+++ y-a-foot-core-services/src/test/java/net/andresbustamante/yafoot/core/services/impl/SportSearchServiceImplTest.java
@@ -26,8 +26,8 @@ class SportSearchServiceImplTest extends AbstractServiceTest {
     void loadSports() throws Exception {
         // Given
         List<Sport> sports = List.of(
-                new Sport(1, SportEnum.FOOTBALL.name()),
-                new Sport(2, SportEnum.RUGBY.name())
+                new Sport((short) 1, SportEnum.FOOTBALL.name()),
+                new Sport((short) 2, SportEnum.RUGBY.name())
         );
 
         // When
